refactor: extract usePointerGestures (pan/zoom/node-drag/marquee core) from SketchCanvas
